requirements.txt
django==1.11.15
numpy==1.15.1
pandas==0.23.4
djangorestframework==3.7.7
markdown==2.6.11
django-filter==1.1.0
six==1.10.0
celery-with-redis==3.0
celery==3.1.25
kombu==3.0.37
billiard==3.3.0.23
django-celery==3.2.2
djangorestframework-jwt==1.11.0
redis==2.10.6
django-rest-auth==0.9.3
dynamic-rest==1.8.1
pillow==5.0.0
django-cors-headers==2.4.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
django-crispy-forms==1.7.2
future==0.16.0
django-import-export==1.0.1
coreapi==2.3.3
httplib2==0.9.2
django-formtools==2.1
